requirements.txt
django==3.2.0
djangorestframework==3.13.0
psycopg2==2.9.3
pytest== 7.1.2
pytest-django==4.5.2